+++ app/src/main/java/ua/com/radiokot/photoprism/features/viewer/di/MediaViewerFeatureModules.kt
@@ -16,6 +16,8 @@ import ua.com.radiokot.photoprism.di.HttpClient
 import ua.com.radiokot.photoprism.di.INTERNAL_DOWNLOADS_DIRECTORY
 import ua.com.radiokot.photoprism.di.VIDEO_CACHE_DIRECTORY
 import ua.com.radiokot.photoprism.env.data.model.EnvSession
+import ua.com.radiokot.photoprism.features.gallery.di.UTC_DATE_TIME_DATE_FORMAT
+import ua.com.radiokot.photoprism.features.gallery.di.UTC_DATE_TIME_YEAR_DATE_FORMAT
 import ua.com.radiokot.photoprism.features.gallery.di.galleryFeatureModules
 import ua.com.radiokot.photoprism.features.viewer.logic.BackgroundMediaFileDownloadManager
 import ua.com.radiokot.photoprism.features.viewer.logic.ThreadPoolBackgroundMediaFileDownloadManager
@@ -58,6 +60,8 @@ val mediaViewerFeatureModules: List<Module> = listOf(
                     externalDownloadsDir = get(named(EXTERNAL_DOWNLOADS_DIRECTORY)),
                     downloadMediaFileViewModel = get(),
                     backgroundMediaFileDownloadManager = get(),
+                    utcDateTimeDateFormat = get(named(UTC_DATE_TIME_DATE_FORMAT)),
+                    utcDateTimeYearDateFormat = get(named(UTC_DATE_TIME_YEAR_DATE_FORMAT)),
                 )
             }